# voice-commerce/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GIMME Voice API",
    "version": "1.0.0",
    "description": "Platform-neutral voice commerce API. Apple, Google and agent adapters call this; it maps onto the same service as the GIMME MCP tools. Amounts are decimal NZD. Every error body has error_code, message, recoverable and suggested_action."
  },
  "servers": [
    {
      "url": "https://voice.gimme.example",
      "description": "Production (placeholder)"
    },
    {
      "url": "http://localhost:8787",
      "description": "Sandbox"
    }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "oauth2",
        "description": "Tokens are issued by GIMME's authorization server (account linking). This API is a resource server only.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://auth.gimme.example/authorize",
            "tokenUrl": "https://auth.gimme.example/token",
            "scopes": {
              "gimme.customer.read": "Read your first name, saved address labels, payment method labels and order preferences",
              "gimme.products.read": "Search GIMME's catalogue",
              "gimme.inventory.read": "Check stock and delivery availability",
              "gimme.cart.create": "Price a basket and prepare an order for your confirmation",
              "gimme.order.create": "Place an order you have confirmed",
              "gimme.order.read": "Read your order status and receipts",
              "gimme.order.cancel": "Cancel an order, where GIMME allows it",
              "gimme.payment.authorize": "Authorize payment for an order you have confirmed, using a saved payment method",
              "gimme.order.confirm": "Record your purchase confirmation (GIMME's own apps only)"
            }
          }
        }
      }
    },
...
  },
  "paths": {
    "/v1/voice/sessions": {
      "post": {
        "summary": "Open a voice session (handshake, identify_customer)",
        "operationId": "postV1VoiceSessions",
        "security": [
          {
            "oauth2": [
              "gimme.customer.read"
            ]
          }
        ],
        "parameters": [
          {
            "name": "GIMME-API-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1.0"
              ]
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "platform": {
                    "type": "string",
                    "enum": [
                      "APPLE",
                      "GOOGLE",
                      "MCP_AGENT",
                      "WEB_AGENT",
                      "SANDBOX"
                    ]
                  },
                  "platform_user_id": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256,
                    "description": "The assistant platform's stable id for this user (must be linked to the GIMME account)."
                  },
                  "platform_session_id": {
                    "type": "string",
                    "maxLength": 256
                  },
                  "conversation_id": {
                    "description": "Platform conversation id; part of the idempotency key for mutations.",
                    "type": "string",
                    "maxLength": 256
                  }
                },
                "required": [
                  "platform",
                  "platform_user_id"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "string"
                    },
                    "customer_id": {
                      "type": "string"
                    },
                    "authenticated": {
                      "type": "boolean"
                    },
                    "customer_name": {
                      "type": "string",
                      "description": "First name only."
                    },
                    "default_address": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "summary": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "summary"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "eligible_to_order": {
                      "type": "boolean"
                    },
                    "eligibility": {
                      "type": "string",
                      "enum": [
                        "ELIGIBLE",
                        "AGE_VERIFICATION_REQUIRED",
                        "IDENTITY_VERIFICATION_REQUIRED",
                        "ACCOUNT_RESTRICTED",
                        "ORDER_NOT_PERMITTED"
                      ]
                    },
                    "expires_at": {
                      "type": "string"
                    },
                    "handshake": {
                      "type": "object",
                      "properties": {
                        "protocol_version": {
                          "type": "string"
                        },
                        "mcp_version": {
                          "type": "string"
                        },
                        "voice_api_version": {
                          "type": "string"
                        },
                        "capability_version": {
                          "type": "string"
                        },
                        "platform": {
                          "type": "string",
                          "enum": [
                            "APPLE",
                            "GOOGLE",
                            "MCP_AGENT",
                            "WEB_AGENT",
                            "SANDBOX"
                          ]
                        },
                        "session_id": {
                          "type": "string"
                        },
                        "customer_id": {
                          "type": "string"
                        },
                        "client_id": {
                          "type": "string"
                        },
                        "scopes": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "request_id": {
                          "type": "string"
                        },
                        "conversation_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "protocol_version",
                        "mcp_version",
                        "voice_api_version",
                        "capability_version",
                        "platform",
                        "session_id",
                        "customer_id",
                        "client_id",
                        "scopes",
                        "request_id"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "session_id",
                    "customer_id",
                    "authenticated",
                    "customer_name",
                    "default_address",
                    "eligible_to_order",
                    "eligibility",
                    "expires_at",
                    "handshake"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/voice/sessions/{id}": {
      "get": {
        "summary": "Read a voice session",
        "operationId": "getV1VoiceSessionsId",
        "security": [
          {
            "oauth2": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "GIMME-API-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1.0"
              ]
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "string"
                    },
                    "customer_id": {
                      "type": "string"
                    },
                    "platform": {
                      "type": "string",
                      "enum": [
                        "APPLE",
                        "GOOGLE",
                        "MCP_AGENT",
                        "WEB_AGENT",
                        "SANDBOX"
                      ]
                    },
                    "platform_session_id": {
                      "type": "string"
                    },
                    "authenticated": {
                      "type": "boolean"
                    },
                    "stage": {
                      "type": "string",
                      "enum": [
                        "DISCOVERY",
                        "CUSTOMER_IDENTIFICATION",
                        "PRODUCT_RESOLUTION",
                        "INVENTORY_VALIDATION",
                        "DELIVERY_VALIDATION",
                        "COMPLIANCE_VALIDATION",
                        "PRICE_CALCULATION",
                        "ORDER_INTENT_CREATED",
                        "CUSTOMER_CONFIRMATION",
                        "PAYMENT_AUTHORIZATION",
                        "FINAL_VALIDATION",
                        "ORDER_CREATION",
                        "ORDER_CONFIRMED",
                        "DELIVERY_TRACKING"
                      ]
                    },
                    "created_at": {
                      "type": "string"
                    },
                    "expires_at": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "session_id",
                    "customer_id",
                    "platform",
                    "authenticated",
                    "stage",
                    "created_at",
                    "expires_at"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/voice/order-intents/{id}/authorize-payment": {
      "post": {
        "summary": "Authorize payment for a confirmed intent",
        "operationId": "postV1VoiceOrderIntentsIdAuthorizePayment",
        "security": [
          {
            "oauth2": [
              "gimme.payment.authorize"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "GIMME-API-Version",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "1.0"
              ]
            }
          },
          {
            "name": "X-Request-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Correlation-Id",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "authorization_token": {
                    "description": "From request_authorization / confirm. Omit only when the customer confirmed out-of-band (hosted confirmation page); the server then uses the authorization it recorded for this intent.",
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 4096
                  },
                  "wallet_token": {
                    "description": "Single-use Apple Pay / Google Pay payment token, when the intent was created with a wallet. Never a card number.",
                    "type": "string",
                    "minLength": 4,
                    "maxLength": 8192
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authorized": {
                      "type": "boolean",
                      "const": true
                    },
                    "payment_authorization_id": {
                      "type": "string"
                    },
                    "payment_reference": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number",
                      "description": "Decimal amount in the given currency, two decimal places."
                    },
                    "currency": {
                      "type": "string",
                      "const": "NZD"
                    }
                  },
                  "required": [
                    "authorized",
                    "payment_authorization_id",
                    "payment_reference",
                    "amount",
                    "currency"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}